view comments in chat_controller and chat_members entity

+++ MessageService/T_03/src/main/java/suraifokkusu/message/chat/entities/Chat.java
@@ -23,6 +23,6 @@ public class Chat {
     @Column(nullable = false, name = "total_messages")
     private Long totalMessages;
 
-    public Chat(int i, String test_lobby, String s, int i1) {
+    public Chat(int i, String test_lobby, String s, Long i1) {
     }
 }
